+++ cli/cli/commands/enclave/enclave.go
@@ -23,10 +23,10 @@ var EnclaveCmd = &cobra.Command{
 }
 
 func init() {
-	EnclaveCmd.AddCommand(ls.LsCmd)
+	EnclaveCmd.AddCommand(ls.EnclaveLsCmd.MustGetCobraCommand())
 	EnclaveCmd.AddCommand(inspect.EnclaveInspectCmd.MustGetCobraCommand())
 	EnclaveCmd.AddCommand(new.NewCmd)
-	EnclaveCmd.AddCommand(stop.StopCmd)
+	EnclaveCmd.AddCommand(stop.EnclaveStopCmd.MustGetCobraCommand())
 	EnclaveCmd.AddCommand(rm.EnclaveRmCmd.MustGetCobraCommand())
-	EnclaveCmd.AddCommand(dump.EnclaveDumpCmd)
+	EnclaveCmd.AddCommand(dump.EnclaveDumpCmd.MustGetCobraCommand())
 }