+++ cli/cli/commands/engine/start/start.go
@@ -5,7 +5,6 @@ import (
 	"fmt"
 	"github.com/kurtosis-tech/kurtosis-cli/cli/command_str_consts"
 	"github.com/kurtosis-tech/kurtosis-cli/cli/defaults"
-	"github.com/kurtosis-tech/kurtosis-cli/cli/helpers/backend_for_cmd"
 	"github.com/kurtosis-tech/kurtosis-cli/cli/helpers/engine_manager"
 	"github.com/kurtosis-tech/kurtosis-cli/cli/helpers/logrus_log_levels"
 	"github.com/kurtosis-tech/stacktrace"
@@ -66,11 +65,15 @@ func run(cmd *cobra.Command, args []string) error {
 		return stacktrace.Propagate(err, "An error occurred parsing log level string '%v'", logLevelStr)
 	}
 
-	kurtosisBackend, err := backend_for_cmd.GetBackendForCmd(WithKubernetes)
-	if err != nil {
-		return stacktrace.Propagate(err, "An error occurred getting a Kurtosis backend")
+	// TODO Hack; remove when we read cluster state from disk
+	var clusterName = "docker"
+	if WithKubernetes {
+		clusterName = "minikube"
 	}
-	engineManager := engine_manager.NewEngineManager(kurtosisBackend)
+	engineManager, err := engine_manager.NewEngineManager(clusterName)
+	if err != nil {
+		return stacktrace.Propagate(err, "An error occurred creating an engine manager connected to cluster '%v'", clusterName)
+	}
 
 	var engineClientCloseFunc func() error
 	var startEngineErr error