+++ cli/cli/helpers/engine_manager/engine_existence_guarantor.go
@@ -8,8 +8,6 @@ import (
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/port_spec"
 	"github.com/kurtosis-tech/kurtosis-cli/cli/command_str_consts"
 	"github.com/kurtosis-tech/kurtosis-cli/cli/helpers/metrics_user_id_store"
-	"github.com/kurtosis-tech/kurtosis-cli/cli/kurtosis_config"
-	"github.com/kurtosis-tech/kurtosis-cli/cli/kurtosis_config/resolved_config"
 	"github.com/kurtosis-tech/kurtosis-engine-api-lib/api/golang/lib/kurtosis_context"
 	"github.com/kurtosis-tech/kurtosis-engine-server/launcher/engine_server_launcher"
 	"github.com/kurtosis-tech/stacktrace"
@@ -37,6 +35,9 @@ type engineExistenceGuarantor struct {
 	// Storing a context in a struct is normally bad, but this visitor is short-lived and behaves more like a function
 	ctx context.Context
 
+	// Whether any engine that gets started should send metrics
+	shouldSendMetrics bool
+
 	// Host machine IP:port of the maybe-started, maybe-not engine (will only be present if the engine status isn't "stopped")
 	preVisitingMaybeHostMachineIpAndPort *hostMachineIpAndPort
 
@@ -63,6 +64,7 @@ func newEngineExistenceGuarantorWithDefaultVersion(
 	ctx context.Context,
 	preVisitingMaybeHostMachineIpAndPort *hostMachineIpAndPort,
 	kurtosisBackend backend_interface.KurtosisBackend,
+	shouldSendMetrics bool,
 	engineServerKurtosisBackendConfigSupplier engine_server_launcher.KurtosisBackendConfigSupplier,
 	logLevel logrus.Level,
 	maybeCurrentlyRunningEngineVersionTag string,
@@ -71,6 +73,7 @@ func newEngineExistenceGuarantorWithDefaultVersion(
 		ctx,
 		preVisitingMaybeHostMachineIpAndPort,
 		kurtosisBackend,
+		shouldSendMetrics,
 		engineServerKurtosisBackendConfigSupplier,
 		defaultEngineImageVersionTag,
 		logLevel,
@@ -82,6 +85,7 @@ func newEngineExistenceGuarantorWithCustomVersion(
 	ctx context.Context,
 	preVisitingMaybeHostMachineIpAndPort *hostMachineIpAndPort,
 	kurtosisBackend backend_interface.KurtosisBackend,
+	shouldSendMetrics bool,
 	engineServerKurtosisBackendConfigSupplier engine_server_launcher.KurtosisBackendConfigSupplier,
 	imageVersionTag string,
 	logLevel logrus.Level,
@@ -97,6 +101,7 @@ func newEngineExistenceGuarantorWithCustomVersion(
 		logLevel:                              logLevel,
 		maybeCurrentlyRunningEngineVersionTag: maybeCurrentlyRunningEngineVersionTag,
 		postVisitingHostMachineIpAndPort:      nil, // Will be filled in upon successful visitation
+		shouldSendMetrics: shouldSendMetrics,
 	}
 }
 
@@ -114,11 +119,6 @@ func (guarantor *engineExistenceGuarantor) VisitStopped() error {
 		return stacktrace.Propagate(err, "An error occurred getting metrics user id")
 	}
 
-	kurtosisConfig, err := getKurtosisConfig()
-	if err != nil {
-		return stacktrace.Propagate(err, "An error occurred getting Kurtosis config")
-	}
-
 	var hostMachineIpAddr net.IP
 	var hostMachinePortSpec *port_spec.PortSpec
 	var engineLaunchErr error
@@ -129,7 +129,7 @@ func (guarantor *engineExistenceGuarantor) VisitStopped() error {
 			kurtosis_context.DefaultKurtosisEngineServerGrpcPortNum,
 			kurtosis_context.DefaultKurtosisEngineServerGrpcProxyPortNum,
 			metricsUserId,
-			kurtosisConfig.GetShouldSendMetrics(),
+			guarantor.shouldSendMetrics,
 			guarantor.engineServerKurtosisBackendConfigSupplier,
 		)
 	} else {
@@ -140,7 +140,7 @@ func (guarantor *engineExistenceGuarantor) VisitStopped() error {
 			kurtosis_context.DefaultKurtosisEngineServerGrpcPortNum,
 			kurtosis_context.DefaultKurtosisEngineServerGrpcProxyPortNum,
 			metricsUserId,
-			kurtosisConfig.GetShouldSendMetrics(),
+			guarantor.shouldSendMetrics,
 			guarantor.engineServerKurtosisBackendConfigSupplier,
 		)
 	}
@@ -242,15 +242,4 @@ func (guarantor *engineExistenceGuarantor) getRunningAndCLIEngineVersions() (*se
 	}
 
 	return runningEngineSemver, launcherEngineSemver, nil
-}
-
-func getKurtosisConfig() (*resolved_config.KurtosisConfig, error) {
-	configStore := kurtosis_config.GetKurtosisConfigStore()
-	configProvider := kurtosis_config.NewKurtosisConfigProvider(configStore)
-
-	kurtosisConfig, err := configProvider.GetOrInitializeConfig()
-	if err != nil {
-		return nil, stacktrace.Propagate(err, "An error occurred getting or initializing config")
-	}
-	return kurtosisConfig, nil
 }