+++ cli/cli/kurtosis_config/resolved_config/kurtosis_config_test.go
@@ -8,29 +8,6 @@ import (
 	"testing"
 )
 
-func TestGetKurtosisBackendSupplier(t *testing.T) {
-	for _, clusterType := range KurtosisClusterTypeValues() {
-		// Set the config appropriately so that we pass validation
-		var config *v1.KubernetesClusterConfigV1
-		switch clusterType {
-		case KurtosisClusterType_Docker:
-			config = nil
-		case KurtosisClusterType_Kubernetes:
-			clusterName := "test"
-			storageClass := "standard"
-			enclaveSizeInMB := uint(10)
-			config = &v1.KubernetesClusterConfigV1{
-				KubernetesClusterName:  &clusterName,
-				StorageClass:           &storageClass,
-				EnclaveSizeInMegabytes: &enclaveSizeInMB,
-			}
-		}
-
-		_, err := getKurtosisBackendSupplier(clusterType, config)
-		require.NoError(t, err)
-	}
-}
-
 func TestNewKurtosisConfigFromRequiredFields_MetricsElectionIsSent(t *testing.T) {
 	config, err := NewKurtosisConfigFromRequiredFields(false)
 	require.NoError(t, err)
@@ -72,4 +49,3 @@ func TestNewKurtosisConfigOverridesAreLatestVersion(t *testing.T) {
 	// check that overrides are actually the latest version
 	require.Equal(t, latestVersion, overrides.ConfigVersion.String())
 }
-