+++ cli/docs/changelog.md
@@ -1,4 +1,17 @@
 # TBD
+### Features
+* `KurtosisClusterConfig` now creates engine container config suppliers
+* Added a config version migration test
+* Connected the config file to use by the commands
+* Added tab-complete to the following commands:
+    * `enclave dump`
+    * `enclave ls`
+    * `enclave rm` 
+    * `service logs`
+    * `service shell`
+
+### Changes
+* The `enclave-size-in-Megabytes` field in Kubernetes cluster field now is optional, and defaults to 10 MB
 
 # 0.17.0
 ### Features